Nvic codegen (#41)
* add system registers for cortex m0 and m1, nvic register generation for the chip

* make warning log a debug log

* usize to u64

* register size is now max 2^16 - 1

+++ tools/regz/src/Peripheral.zig
@@ -3,7 +3,7 @@ const std = @import("std");
 name: []const u8,
 version: ?[]const u8,
 description: ?[]const u8,
-base_addr: ?usize,
+base_addr: ?u64,
 
 pub fn format(
     peripheral: @This(),
